refactor: modularize ml sync actions
